# dropwizard-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.7.1",
        "title": "Red Hat Massages Server"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
        "/clients": {
            "get": {
                "description": "GETs all {@link Client}s that can be found.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "headers": {
                        },
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Client"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Updates a {@link Client} to an updated value.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Client"
                        },
                        "description": "updated {@link Client}"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "headers": {
                        },
                        "schema": {
                            "$ref": "#/definitions/Client"
                        }
                    },
                    "204": {
                        "description": "No Content",
                        "headers": {
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "headers": {
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "headers": {
                        }
                    }
                }
            }
        },
...
        "/clients/retrieve-info": {
            "get": {
                "description": "GETs local information of a {@link Client} (currently only his subscription value). For new\n {@link User}s creates their {@link Client} representation while returning {@link User}s get\n their {@link Client} representation updated if a change is detected.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User"
                        },
                        "description": "authenticated {@link User}"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "headers": {
                        },
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "headers": {
                        }
                    }
                }
            }
        },
...
        "/clients/{sub}": {
            "delete": {
                "description": "Deletes a {@link Client} given by the subject. All massages assigned to that client are freed\n before the deletion or, if the client is a masseur, those massages are automatically removed.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "name": "sub",
                        "in": "path",
                        "required": true,
                        "description": "{@link Client} subject"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "headers": {
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "headers": {
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Client": {
            "properties": {
                "email": {
                    "type": "string"
                },
                "masseur": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "sub": {
                    "type": "string"
                },
                "subscribed": {
                    "type": "boolean"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
...
        "User": {
            "properties": {
                "admin": {
                    "type": "boolean"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "masseur": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "subject": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        }
    }
}